Refactor DTOs; Remove unit tests for tag-entity for now
