--- backend/src/tags/dto/create-tag.dto.ts
@@ -1,17 +1,0 @@
-import { IsHexColor, IsString, MaxLength } from "class-validator";
-
-class CreateTagDto {
-  @IsString()
-  @MaxLength(50, { message: "Tag must be less then 50 chars" })
-  title: string;
-
-  @IsString()
-  @IsHexColor({ message: "Text color must be HEX like #RRGGBB or #RGB" })
-  textColor: string;
-
-  @IsString()
-  @IsHexColor({ message: "Text color must be HEX like #RRGGBB or #RGB" })
-  backgroundColor: string;
-}
-
-export { CreateTagDto };
