--- backend/src/tags/dto/update-tag.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/mapped-types';
-import { CreateTagDto } from './create-tag.dto';
-
-export class UpdateTagDto extends PartialType(CreateTagDto) {}
