--- backend/src/tags/tags.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { TagsController } from './tags.controller';
-import { TagsService } from './tags.service';
-
-describe('TagsController', () => {
-  let controller: TagsController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [TagsController],
-      providers: [TagsService],
-    }).compile();
-
-    controller = module.get<TagsController>(TagsController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
