--- backend/src/tags/tags.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { TagsService } from './tags.service';
-
-describe('TagsService', () => {
-  let service: TagsService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [TagsService],
-    }).compile();
-
-    service = module.get<TagsService>(TagsService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
